Migrated DeleteLayerHandler
